==> 99_backup.ipynb <==
from sklearn.ensemble import RandomForestClassifier

model = (RandomForestClassifier(random_state=42), 'Random Forest')

# Random Forest
rf_model, rf_model_name = RandomForestClassifier(random_state=42), 'Random Forest'
rf_model, rf_model_name, rf_results_df, rf_fold_results = train_and_find_best_model(rf_model, rf_model_name, X, y, metric='Macro F1')
rf_model_path = save_best_model(rf_model, rf_model_name)
# ROS
rf_model_ros, rf_model_name_ros = RandomForestClassifier(random_state=42), 'Random Forest ROS'
rf_model_ros, rf_model_name_ros, rf_results_ros_df, rf_fold_results_ros = train_and_find_best_model(rf_model_ros, rf_model_name_ros, X_ros, y_ros, metric='Macro F1')
rf_model_ros_path = save_best_model(rf_model_ros, rf_model_name_ros)
# SMOTE
rf_model_smote, rf_model_name_smote = RandomForestClassifier(random_state=42), 'Random Forest SMOTE'
rf_model_smote, rf_model_name_smote, rf_results_smote_df, rf_fold_results_smote = train_and_find_best_model(rf_model_smote, rf_model_name_smote, X_smote, y_smote, metric='Macro F1')
rf_model_smote_path = save_best_model(rf_model_smote, rf_model_name_smote)

rf_results = pd.concat([rf_results_df, rf_results_ros_df, rf_results_smote_df], axis=0)

pd.DataFrame(rf_fold_results), pd.DataFrame(rf_fold_results_ros), pd.DataFrame(rf_fold_results_smote), 

rf_model = joblib.load(rf_model_path)
rf_model_ros = joblib.load(rf_model_ros_path)
rf_model_smote = joblib.load(rf_model_smote_path)

rf_results_test = evaluate_model_on_test_data(rf_model, feature, target, 'Random Forest')
rf_results_ros_test = evaluate_model_on_test_data(rf_model_ros, feature, target, 'Random Forest ROS')
rf_results_smote_test = evaluate_model_on_test_data(rf_model_smote, feature, target, 'Random Forest SMOTE')

rf_results_test, rf_results_ros_test, rf_results_smote_test,
